% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/englamaps.R
\docType{data}
\name{aurnsites}
\alias{aurnsites}
\title{AURN data sites (123) in England}
\format{
A data frame with 123 rows and 9 columns:
\describe{
  \item{id}{Internal: The id number of the sites.}
   \item{SiteName}{Name of the site}
   \item{UK.AIRID}{UK Air id Code}
   \item{EnvironmentType}{Environment type of the site}
   \item{Latitude}{Latitude}
   \item{Longitude}{Longitude}
   \item{Northing}{Gives the UTM-Y coordinate of the site}
  \item{Easting}{Gives the UTM-X coordinate of the site}
  \item{Altitudeinm}{Altitude in metre}
}
}
\usage{
aurnsites
}
\description{
AURN data sites (123) in England
}
\examples{
\dontrun{
 head(aurnsites)
}
}
\keyword{datasets}
