% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{colpalette}
\alias{colpalette}
\title{The color palette used to draw maps in the book
\insertCite{Sahubook;textual}{databmstdr}}
\format{
An object of class \code{character} of length 5.
}
\usage{
colpalette
}
\description{
The color palette used to draw maps in the book
\insertCite{Sahubook;textual}{databmstdr}
}
\examples{
\dontrun{
 colpalette
}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
