% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/englamaps.R
\docType{data}
\name{eaststates}
\alias{eaststates}
\title{Names of the 42 states and regions in the eastern USA used to draw the maps of the
eastern US.}
\format{
An object of class \code{character} of length 42.
}
\source{
Sahu (2021)
}
\usage{
eaststates
}
\description{
Names of the 42 states and regions in the eastern USA used to draw the maps of the
eastern US.
}
\examples{
\dontrun{
 eaststates
 library(ggplot2)
 useastmap <- map_data(database="state",regions=eaststates)
 head(useastmap)
 p <-  ggplot(data=useastmap, aes(x=long, y=lat, group = group)) +
 geom_polygon(colour='black',size=0.8, fill=NA) +
 theme(axis.text.x = element_blank(), axis.text.y = element_blank(), axis.ticks = element_blank()) +
 ggsn::north(data=useastmap, location="bottomright", symbol=12) +
 ggsn::scalebar(data =useastmap, dist =500, location ="bottomleft", transform=T, dist_unit = "km",
 st.dist = .05, st.size = 5, height = .06, st.bottom=T, model="WGS84")
 p
}
}
\keyword{datasets}
