% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{englainfo}
\alias{englainfo}
\title{Data frame providing information regarding the 313 local authorities
in England.}
\format{
An object of class \code{data.frame} with 313 rows and 8 columns.
}
\source{
\insertCite{Sahubook;textual}{databmstdr}

 @format A data frame with 313 rows and 8 columns:
\describe{
  \item{Areacode}{Codes of the local authorities}
  \item{id}{Internal: The id number of the polygons.}
   \item{Areaname}{Name of the local authority}
   \item{Regioncode}{Codes of the regions}
   \item{Rid}{Internal: The id number of the region.}
   \item{Regionname}{Name of the regions}
  \item{centreEasting}{Gives the UTM-X coordinate of the centroid of the Areacode}
  \item{centreNorthing}{Gives the UTM-Y coordinate of the centroid of the Areacode}
}
}
\usage{
englainfo
}
\description{
Data frame providing information regarding the 313 local authorities
in England.
}
\examples{
\dontrun{
 head(englainfo)
}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
