% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/englamaps.R
\docType{data}
\name{englamap}
\alias{englamap}
\title{Data frame providing the polygons for plotting the map of
local authorities in England}
\format{
An object of class \code{data.frame} with 258188 rows and 8 columns.
}
\source{
Sahu and Bohning (2021)
 @format A data frame with 258188 rows and 7 columns:
\describe{
  \item{id}{Internal: The id number of the polygons.}
  \item{long}{Gives the UTM-X coordinates}
  \item{lat}{Gives the UTM-Y coordinates}
  \item{order}{Internal variable: order of the polygons}
  \item{hole}{Internal variable: binary variable}
  \item{piece}{Internal variable}
  \item{group}{Internal variable: group variable for plotting}
  \item{Areacode}{Areacode of the local authorties}
}
}
\usage{
englamap
}
\description{
Data frame providing the polygons for plotting the map of
local authorities in England
}
\examples{
\dontrun{
 head(englamap)
}
}
\keyword{datasets}
