% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{engregmap}
\alias{engregmap}
\title{Data frame providing the polygons for plotting a map of the
administrative regions in England}
\format{
An object of class \code{data.frame} with 916012 rows and 9 columns.
}
\source{
\insertCite{Sahubook;textual}{databmstdr}

 @format A data frame with 258188 rows and 7 columns:
\describe{
  \item{id}{Internal: The id number of the polygons.}
  \item{long}{Gives the UTM-X coordinates}
  \item{lat}{Gives the UTM-Y coordinates}
  \item{order}{Internal variable: order of the polygons}
  \item{hole}{Internal variable: binary variable}
  \item{piece}{Internal variable}
  \item{group}{Internal variable: group variable for plotting}
  \item{Region}{Name of the regions}
  \item{Regioncode}{Codes of the regions}
}
}
\usage{
engregmap
}
\description{
Data frame providing the polygons for plotting a map of the
administrative regions in England
}
\examples{
\dontrun{
 head(engregmap)
}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
