% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{kdat}
\alias{kdat}
\title{Data frame containing the Kenya vaccination data}
\format{
A data frame with 47 rows and 9 columns:
\describe{#
  \item{id}{id number of the 47 counties in Kenya}
  \item{countyname}{Name of the county}
   \item{n}{Number of children sampled}
   \item{yvac}{Number of children vaccinated }
   \item{propprimary}{proportion of women aged 15-49 years
   who had completed primary education}
   \item{traveltime}{travel time to the nearest health facility, which are
    county averages of 5 by 5 kilometer grid level values}
   \item{nightlight}{intensity which are also county averages of the corresponding
   5 by 5 kilometer  grid level values.}
   \item{wted_n}{Weighted number of children sampled}
   \item{wted_yvac}{Weighted number of children vaccinated}
}
}
\source{
\insertCite{Sahubook;textual}{databmstdr}
\insertCite{Utazi2021;textual}{databmstdr}
}
\usage{
kdat
}
\description{
Data frame containing the Kenya vaccination data
}
\examples{
\dontrun{
 summary(kdat)

}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
