% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{materncov}
\alias{materncov}
\title{Banerjee et al Matern covariance function}
\usage{
materncov(d, phi, nu)
}
\arguments{
\item{d}{is the distance}

\item{phi}{is the rate of decay}

\item{nu}{is the smoothness parameter}
}
\value{
Returns the Matern covariance for distance object d
}
\description{
Banerjee et al Matern covariance function
}
