% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/englamaps.R
\docType{data}
\name{colpalette}
\alias{colpalette}
\title{The color palette used to draw most of the maps}
\format{
An object of class \code{character} of length 5.
}
\source{
Sahu and Bohning (2021)
}
\usage{
colpalette
}
\description{
The color palette used to draw most of the maps
}
\examples{
\dontrun{
 colpalette
}
}
\keyword{datasets}
